app/requirements.txt
streamlit==1.31.0
elasticsearch==8.9.0
sentence-transformers==2.7.0
huggingface-hub==0.23.0
openai==1.30.0
httpx==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
torch==2.2.0
numpy<2.0.0